backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Async task queue
celery[redis]==5.3.6
redis==5.0.1

# AI APIs
google-generativeai==0.3.2
faster-whisper==1.0.3

# Core dependencies
numpy==1.24.3

# Image/Video processing
Pillow==10.2.0
opencv-python-headless==4.8.1.78

# Audio processing
librosa==0.10.1
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
